Refactor to multiple steps
