+++ src/okta/types/devices.ts
@@ -1,0 +1,11 @@
+export interface OktaFactor {
+  id: string;
+  factorType: string;
+  provider: string;
+  vendorName?: string;
+  device?: string;
+  deviceType?: string;
+  status: string;
+  created: string;
+  lastUpdated: string;
+}
