+++ src/okta/types/index.ts
@@ -1,0 +1,13 @@
+export * from "./applications";
+export * from "./client";
+export * from "./devices";
+export * from "./groups";
+export * from "./users";
+
+/**
+ * Account information derived from the domain/url.
+ */
+export interface OktaAccountInfo {
+  name: string;
+  preview: boolean;
+}
